[core] refactor callApi into multiple pieces, use jest over sinon for spys
