+++ packages/superset-ui-core/src/callApi/rejectAfterTimeout.js
@@ -1,0 +1,13 @@
+// returns a Promise that rejects after the specified timeout
+export default function rejectAfterTimeout(timeout) {
+  return new Promise((resolve, reject) => {
+    setTimeout(
+      () =>
+        reject({
+          error: 'Request timed out',
+          statusText: 'timeout',
+        }),
+      timeout,
+    );
+  });
+}
