+++ packages/superset-ui-core/test/utils/throwIfCalled.js
@@ -1,0 +1,3 @@
+export default function throwIfCalled() {
+  throw Error('Unexpected call to throwIfCalled()');
+}
